+++ Assets/Scripts/Gameplay/AI/StateMachine/IAIState.cs
@@ -6,16 +6,4 @@ namespace Gameplay.AI
     {
         Task<StateResult> Launch();
     }
-    
-    public class StateResult
-    {
-        public IAIState NextState { get; }
-        public bool IsFinished { get; }
-
-        public StateResult(IAIState nextState, bool isFinished)
-        {
-            NextState = nextState;
-            IsFinished = isFinished;
-        }
-    }
 }
